Update structures.h
logic error

+++ structures.h
@@ -9,7 +9,7 @@ struct player{
 	int hand[2][2]; // storing hands 
 	double chipsput; // stroing deadn chips in front
 	bool allin; // true indicates the player has all in
-	int sidepool; // only have value when a player all in and determine the pot he can get if he wins
+	double sidepool; // only have value when a player all in and determine the pot he can get if he wins
 	int value; // assigning a value of the rank of player after check win 
 	player *next;   //circular linked list
 };
